refactor: token.hを作成
