+++ srcs/lexer/token_utils.c
@@ -6,7 +6,7 @@
 /*   By: totaisei <totaisei@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/15 17:42:00 by totaisei          #+#    #+#             */
-/*   Updated: 2021/02/16 19:40:58 by totaisei         ###   ########.fr       */
+/*   Updated: 2021/02/18 15:55:46 by totaisei         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -80,19 +80,4 @@ t_token			*token_init(size_t len, t_token *prev)
 	res->next = NULL;
 	res->prev = prev;
 	return (res);
-}
-
-void			token_add_front(t_tokeniser *toker)
-{
-	t_token *tmp_token;
-
-	if (toker->tok_i > 0)
-	{
-		toker->token->data[toker->tok_i] = '\0';
-		tmp_token =
-		token_init(toker->str_len - toker->str_i, toker->token);
-		toker->token->next = tmp_token;
-		toker->token = tmp_token;
-		toker->tok_i = 0;
-	}
 }
